requirements.txt
fpdf==2.5.6
Pillow==9.1.1